upload

+++ api/middleware/upload.js
@@ -6,8 +6,8 @@ cloudinary.config({
     api_secret: process.env.CLOUND_API_SECRET,
 });
 
-const upload = async (req, res, next) => {
-    console.log(req.files)
+const upload = async(req, res, next) => {
+    console.log(req.body);
     if (!req.files)
         return res.status(400).json({ success: false, message: "Khong co file" });
     var files;
@@ -32,11 +32,12 @@ const upload = async (req, res, next) => {
                 message: "Dung lượng file" + file.name + " quá lớn",
             });
     });
+
     req.results = [];
     for (let i = 0; i < files.length; i++) {
         try {
             files[i].tempFilePath = "./tmp/" + Date.now();
-            fs.writeFile(files[i].tempFilePath, files[i].data, function (err) {
+            await fs.writeFile(files[i].tempFilePath, files[i].data, function(err) {
                 if (err) {
                     console.log(err);
                     return res.status(400).json({
@@ -45,31 +46,35 @@ const upload = async (req, res, next) => {
                     });
                 }
             });
+
             await cloudinary.v2.uploader.upload(
                 files[i].tempFilePath, { folder: req.body.folder },
-                function (error, result) {
-                    req.results.push({ url: result.url, public_id: result.public_id });
-                    fs.unlink(files[i].tempFilePath, (err) => {
-                        if (err)
-                            return res.status(400).json({
-                                success: false,
-                                message: "Xảy ra lỗi trong quá trình tải hình ảnh!",
-                            });
-                    });
+                function(error, result) {
+                    if (error) console.log(error);
+                    if (result) {
+                        req.results.push({ url: result.url, public_id: result.public_id });
+                        fs.unlink(files[i].tempFilePath, (err) => {
+                            if (err)
+                                return res.status(400).json({
+                                    success: false,
+                                    message: "Xảy ra lỗi trong quá trình tải hình ảnh!",
+                                });
+                        });
+                    }
                 }
             );
         } catch (err) {
             for (let i = 0; i < req.results.length; i++) {
                 await unlink(req.results[i].public_id);
             }
-
             req.results = [];
         }
     }
+
     next();
 };
 
-const unlink = async (public_id) => {
+const unlink = async(public_id) => {
     try {
         let rel;
         await cloudinary.v2.uploader.destroy(public_id, (err, result) => {
